+++ tsconfig.json
@@ -2,9 +2,6 @@
   // https://nuxt.com/docs/guide/concepts/typescript
   "extends": "./.nuxt/tsconfig.json",
   "compilerOptions": {
-    "types": [
-      "vitest/importMeta"
-    ],
     "plugins": [
       {
         "name": "@effect/language-service"
