+++ vitest.config.ts
@@ -2,9 +2,6 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    includeSource: ['server/**/*.ts', 'features/**/*.ts'],
-  },
-  define: {
-    'import.meta.vitest': 'undefined',
-  },
+    includeSource: ['server/**/*.test.ts', 'features/**/*.test.ts'],
+  }
 })
